+++ src/plugins/markdown.js
@@ -3,6 +3,8 @@
 const through = require('through2');
 const MarkdownIt = require('markdown-it');
 const hljs = require('highlight.js');
+const attrs = require('markdown-it-attrs');
+const container = require('markdown-it-container');
 const defaults = {
     html: true,
     linkify: true,
@@ -24,6 +26,19 @@ module.exports = function (options) {
         md = options;
     } else {
         md = new MarkdownIt(defaults);
+
+        ['section', 'figure', 'figcaption', 'header', 'footer'].forEach(name => {
+            md.use(container, name, {
+                validate: params => params.trim() === name || params.trim().startsWith(`${name} `),
+                render: (tokens, idx, _options, env, self) => {
+                    tokens[idx].tag = name;
+                    return self.renderToken(tokens, idx, _options, env, self);
+                }
+            });
+        });
+
+        md.use(container, 'div');
+        md.use(attrs);
 
         if (typeof options === 'function') {
             options(md);