+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## UNRELEASED
+
+### Added
+
+* Included `markdown-it-container` plugin by default
+* Included `markdown-it-attrs` plugin by default
+
 ## [0.13.0] - 2017-08-10
 
 ### Changed